+++ CHANGELOG.md
@@ -2,7 +2,8 @@
 
 ## 0.0.4 - WIP
 ### Add
-- *of()* static method, for creating Arr instance from parameters
+- *of()* static method, for creating Arr instance from parameters;
+- *keys()* method for extracting keys from Arr;
 - *sum()* method;
 - *avg()* method;
 - *doc/arr.md*: class usage documentation.