+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## 0.0.4 - WIP
+## 0.0.4 - 2021-12-28
 ### Add
 - *of()* static method, for creating Arr instance from parameters;
 - *keys()* method for extracting keys from Arr;